Changes based on code review.
