+++ src/AltinnCore/Common/Languages/ini/en.ini
@@ -12,6 +12,7 @@ create_new = Create new
 delete = Delete
 disabled = Disabled
 enabled = Enabled
+for = for
 label = Label
 no_options = No options available
 required = Required
