+++ src/AltinnCore/Common/Languages/ini/nb.ini
@@ -12,6 +12,7 @@ create_new = Opprett ny
 delete = Slett
 disabled = Deaktivert
 enabled = Aktivert
+for = for
 label = Navn
 no_options = Ingen alternativer tilgjenglig
 required = Obligatorisk
